Organised all the files
